Compyle on windows

+++ tests/tst_parser.h
@@ -76,7 +76,7 @@ TEST( JsonioParser, TestDouble )
 
 TEST( JsonioParser, TestInt )
 {
-    auto TEST_INT = [](const std::string & json_string, const long expected)
+    auto TEST_INT = [](const std::string & json_string, const long long expected)
     {
         auto jsFree = json::loads(json_string);
         EXPECT_EQ( JsonBase::Int, jsFree[0].type() );
@@ -91,8 +91,8 @@ TEST( JsonioParser, TestInt )
     TEST_INT("[1000000]", 1000000);
 
     // limits
-   TEST_INT("[-9223372036854775807]", -9223372036854775807);
-   TEST_INT("[+9223372036854775807]", +9223372036854775807);
+   TEST_INT("[-9223372036854775807]", -9223372036854775807L);
+   TEST_INT("[+9223372036854775807]", +9223372036854775807L);
 }
